move pipeline stages to subdirectory
