+++ examples/ramen_rec/app/candidate_generators/similar_ramen_candidate_generator.py
@@ -1,4 +1,4 @@
-from frex.candidate_generators import CandidateGenerator
+from frex.pipeline_stages.candidate_generators import CandidateGenerator
 from typing import Dict, List, Generator
 from rdflib import URIRef
 from frex.models import Explanation, Candidate
