+++ examples/ramen_rec/app/scorers/ramen_style_scorer.py
@@ -1,4 +1,4 @@
-from frex.scorers import CandidateScorer
+from frex.pipeline_stages.scorers import CandidateScorer
 from examples.ramen_rec.app.models.ramen_context import RamenContext
 from examples.ramen_rec.app.models.ramen_candidate import RamenCandidate
