+++ examples/ramen_rec/tests/test_ramen_pipelines.py
@@ -1,5 +1,5 @@
 from frex.pipelines import PipelineExecutor
-from frex.scorers import CandidateRanker
+from frex.pipeline_stages.scorers import CandidateRanker
 from examples.ramen_rec.app.models import RamenContext
 from examples.ramen_rec.tests.conftest import placeholder_ramen_candidate
