+++ frex/__init__.py
@@ -1,8 +1,6 @@
-from ._pipeline_stage import _PipelineStage
-from ._pipeline import _Pipeline
 from .models import *
-from .candidate_generators import *
-from .filters import *
-from .scorers import *
+from frex.pipeline_stages.candidate_generators import *
+from frex.pipeline_stages.filters import *
+from frex.pipeline_stages.scorers import *
 from .stores import *
 from .pipelines import *
