+++ frex/pipeline_stages/_pipeline_stage.py
