+++ frex/pipeline_stages/filters/candidate_filterer.py
@@ -1,7 +1,7 @@
 from abc import abstractmethod
 from typing import Generator
 from frex.models import Explanation, Candidate, Context
-from frex import _PipelineStage
+from frex.pipeline_stages import _PipelineStage
 
 
 class CandidateFilterer(_PipelineStage):
