+++ frex/pipeline_stages/scorers/candidate_scorer.py
@@ -1,7 +1,7 @@
 from abc import abstractmethod
 from typing import Tuple, Generator
 from frex.models import Explanation, Candidate, Context
-from frex import _PipelineStage
+from frex.pipeline_stages import _PipelineStage
 
 
 class CandidateScorer(_PipelineStage):
