+++ frex/pipelines/__init__.py
@@ -1,1 +1,2 @@
+from ._pipeline import _Pipeline
 from .pipeline_executor import PipelineExecutor
